Apply thresholding to the image and documenting the process
In this project, Image is read and thresholding algorithm is applied which
highlights impoertant pixels. Different thresholding value is used to obtain
multiple output. Grayscale Image is at first blurred and edge detection algorithm
is applied. After that thresholding is performed. The steps used in this process
are documented in README.md file