Apply SIFT algorithm to detect keypoints and its orientation
SIFT algorithm was applied to the image in order to obtain the feature and
its orientation. Gaussian Blur was used to blur the image to different scales
and Images were decrease in Size which gives their respective octaves. The
blurred image were subtracted to obtain Difference of Gaussian and the this
difference was used to calculate keypoints and orientation